Removal of Attack phase and including in Player Class
